requirements.txt
imapclient
pyyaml
argon2_cffi
pycryptodome
protobuf
XEdDSA
DoubleRatchet
pynacl

pytest
pytest-timeout
pytest-cov
faker
codecov
